% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asimon.R
\name{asimon}
\alias{asimon}
\title{Calculates parameters for adaptive two-stage Simon's design for clinical trials.}
\usage{
asimon(a_max, b1_max, b2_max, p0, p1, p2)
}
\arguments{
\item{a_max}{type I error constraint.}

\item{b1_max, b2_max}{type II error constraints for stages 1 and 2.}

\item{p0}{null hypothesis response rate.}

\item{p1, p2}{alternative hypothesis rates.}
}
\value{
List of 4 elements for each type of optimality.
}
\description{
Calculates parameters for adaptive two-stage Simon's design for clinical trials.
}
\examples{
asimon(0.05, 0.2, 0.1, 0.05, 0.2, 0.25)
}
